chore: fix other minor layout issues
